lot's of small and big performance improvements in the new arithmetics
